% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{alignStrands}
\alias{alignStrands}
\title{Aligns strands between 2 summary statistics data}
\usage{
alignStrands(
  inputData,
  A1.x = "A1.x",
  A2.x = "A2.x",
  A1.y = "A1.y",
  A2.y = "A2.y"
)
}
\arguments{
\item{inputData}{dataframe of both studies with columns:  A1.x, A2.x, A1.y, A2.y}

\item{A1.x}{column for A1 allele in study 1}

\item{A2.x}{column for A2 allele in study 1}

\item{A1.y}{column for A1 allele in study 2}

\item{A2.y}{column for A1 allele in study 2}
}
\value{
returns real value of the approximate correlation
}
\description{
Aligns strands between 2 summary statistics data
}
