% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{convertDSCToDense}
\alias{convertDSCToDense}
\title{Convert DSC to Dense matrix}
\usage{
convertDSCToDense(pop1LDmatrix, numparts = 3)
}
\arguments{
\item{pop1LDmatrix}{LD reference matrix in RDS (dsCMatrix) format for target population}

\item{numparts}{(optional) how many parts should be used for converting the matrix, if memory becomes an issue use higher numbers}
}
\value{
returns a dense LD matrix
}
\description{
Helper function that converts an LD from sparse DSC to Dense format, in a given number of parts (to overcome RAM limitations)
}
