% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{shaPRS_adjust}
\alias{shaPRS_adjust}
\title{Create lFDR corrected Q-test statistics for each SNP}
\usage{
shaPRS_adjust(inputData, rho = 0, thresholds = vector())
}
\arguments{
\item{inputData}{summary statistics table that has header with the following columns: SNP    CHR    BP    Beta_A    SE_A    Beta_B    SE_B}

\item{rho}{overlap between studies. 0 for no overlap and 1 for complete overlap. default: 0}

\item{thresholds}{vector of thresholds to be used to create list of SNPs (default empty)}
}
\value{
returns object with two fields, (1) lFDRTable: a 2 column file with the following signature SNPID lFDR (2) hardThresholds list of SNPids that failed the heterogeneity test at each threshold
}
\description{
it performs:
(1) modified Cochran's Q-test which optionally adjusts for overlapping controls
(2) lFDR estimation on the p-values from the above Cochran's Q-test
(3) lists SNPs that are fail the heterogeneity test at specified thresholds (optional)
}
\examples{
inputDataLoc <- system.file("extdata", "shapersToydata.txt", package = "shaPRS")
inputData= read.table(inputDataLoc, header = TRUE)
results = shaPRS_adjust(inputData, thresholds=c(0.5,0.99))

}
