% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{shaPRS_blend}
\alias{shaPRS_blend}
\title{Blend summary statistics according to a continuous weighting scheme}
\usage{
shaPRS_blend(subpheno, CombinedPheno, blendingFactors)
}
\arguments{
\item{subpheno}{Subphenotype LDPred formatted GWAS summary statistics table  that has header with the following columns: chr    pos    SNP    A1    A2    Freq1.Hapmap    b    se    p    N}

\item{CombinedPheno}{Combined phenotype LDPred formatted GWAS summary statistics table, same signature}

\item{blendingFactors}{a headerless 2 column table of the SNPid and lFDR, (produced by shaPRS_adjust)}
}
\value{
returns an LDPred formatted summary statistics table
}
\description{
This function continuously blends the sub-phenotype and the combined phenotype summary statistics
and generates an LDPred formatted table.
}
\examples{
subphenoLoc <- system.file("extdata", "phenoA_sumstats", package = "shaPRS")
CombinedPhenoLoc <- system.file("extdata", "Combined_sumstats", package = "shaPRS")
blendFactorLoc <- system.file("extdata", "myOutput_SNP_lFDR", package = "shaPRS")
subpheno= read.table(subphenoLoc, header = TRUE)
CombinedPheno= read.table(CombinedPhenoLoc, header = TRUE)
blendingFactors= read.table(blendFactorLoc, header = FALSE)
blendedSumstats = shaPRS_blend(subpheno, CombinedPheno, blendingFactors)

}
