% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{shaPRS_composite}
\alias{shaPRS_composite}
\title{Composite shaPRS: produce summary statistics according to hard thresholds}
\usage{
shaPRS_composite(subpheno, CombinedPheno, hard_threshold_results)
}
\arguments{
\item{subpheno}{Subphenotype LDPred formatted GWAS summary statistics table  that has header with the following columns: chr    pos    SNP    A1    A2    Freq1.Hapmap    b    se    p    N}

\item{CombinedPheno}{Combined phenotype LDPred formatted GWAS summary statistics table, same signature}

\item{hard_threshold_results}{a list of hard thresholds, in the format produced by shaPRS_adjust()}
}
\value{
returns a list of LDPred formatted summary statistics tables, one for each threshold
}
\description{
Loops through a list of hard thresholds, and swaps between the Combined and Subphenotype estimates
based on the list of SNPs found to be heterogenous for a given threshold. This process would swap in
the subphenotype estimates for the SNPs found to be heterogeneous.
}
\examples{
inputDataLoc <- system.file("extdata", "shapersToydata.txt", package = "shaPRS")
inputData= read.table(inputDataLoc, header = TRUE)
results = shaPRS_adjust(inputData, thresholds=c(0.5,0.99))
subphenoLoc <- system.file("extdata", "phenoA_sumstats", package = "shaPRS")
CombinedPhenoLoc <- system.file("extdata", "Combined_sumstats", package = "shaPRS")
subpheno= read.table(subphenoLoc, header = TRUE)
CombinedPheno= read.table(CombinedPhenoLoc, header = TRUE)
CompositeSumstats = shaPRS_composite(subpheno, CombinedPheno, results$hardThresholds)

}
