% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{shaPRS_LDGen}
\alias{shaPRS_LDGen}
\title{Convenience function to generate a shaPRS specific LD reference panel for cross-ancestry analyses}
\usage{
shaPRS_LDGen(
  Pop1LDRefLoc,
  Pop2LDRefLoc,
  blendFactorLoc,
  adjustinputLoc,
  outputLoc,
  discardAmbiguousSNPs = F,
  memoryEfficiency = 5
)
}
\arguments{
\item{Pop1LDRefLoc}{Location of the folder of the LDpred2 formatted LD-reference matrices for the 22 autosomes together with a map.rds file for the proximal study}

\item{Pop2LDRefLoc}{Location of the folder of the LDpred2 formatted LD-reference matrices for the 22 autosomes together with a map.rds file for the adjunct study}

\item{blendFactorLoc}{Location for the lFDR data file produced by shaPRS(), postfix: "_SNP_lFDR"}

\item{adjustinputLoc}{Location for the file produced by the shaPRS(), postfix: "_adjustinput"}

\item{outputLoc}{Output location}

\item{discardAmbiguousSNPs}{(optional) if ambiguous SNPs (G/C and A/T) should be discarded (default TRUE)}

\item{memoryEfficiency}{(optional) larger numbers result in longer runs but lower memory usage (default 5)}
}
\description{
Wrapper function that loads and processes the LD data for two populations, aligns it with summary data for shaPRS and then generate a full new LD-ref panel for 22 autosomes (this should be used instead of LDRefBlend)
}
\examples{
Pop1LDRefLoc <- paste0(system.file("extdata", "", package = "shaPRS"), "/")
Pop2LDRefLoc <- paste0(system.file("extdata", "", package = "shaPRS"), "/")
blendFactorLoc <- system.file("extdata", "pop_SNP_lFDR", package = "shaPRS")
adjustinputLoc <- system.file("extdata", "pop_adjustinput", package = "shaPRS")
outputLoc <- "<YOUR LOCATION>"
# shaPRS_LDGen(Pop1LDRefLoc, Pop2LDRefLoc, blendFactorLoc, adjustinputLoc, outputLoc)

}
