% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{shaPRS_rho}
\alias{shaPRS_rho}
\title{Calculate rho to be used in shaPRS_adjust()}
\usage{
shaPRS_rho(nkl0, nk1, nk0, nl1, nl0)
}
\arguments{
\item{nkl0}{number of controls overlapping between studies}

\item{nk1}{number of cases in study k}

\item{nk0}{number of controls in study k}

\item{nl1}{number of cases in study l}

\item{nl0}{number of controls in study l}
}
\value{
returns real value of the approximate correlation
}
\description{
Convenience function to estimate the correlation between  two studies with overlapping controls
for more details see:
Lin et al. 'Meta-Analysis of Genome-wide Association Studies with Overlapping Subjects' (2009
ncbi.nlm.nih.gov/pmc/articles/PMC2790578
}
\examples{
rho = shaPRS_rho(nkl0 = 9492,nk1 = 3810, nk0= 9492, nl1= 3765,nl0= 9492)

}
