% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{inverse_metaAnalaysis}
\alias{inverse_metaAnalaysis}
\title{Generic inverse variance meta analysis}
\usage{
inverse_metaAnalaysis(proximal, adjunct, rho = 0)
}
\arguments{
\item{proximal}{dataframe for main proximal}

\item{adjunct}{dataframe for other proximal}

\item{rho}{(optional) overlap between studies}
}
\value{
returns Combinedpheno dataframe that can be plugged into shaPRS_blend or shaPRS_composite
}
\description{
Convenience function to produce the combined phenotype estimate if we only have summary stats for phenoA and pheno B
}
