% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{inverse_metaAnalaysis}
\alias{inverse_metaAnalaysis}
\title{Generic inverse variance meta analysis}
\usage{
inverse_metaAnalaysis(subpheno, subpheno_other, rho = 0)
}
\arguments{
\item{subpheno}{dataframe for main subpheno}

\item{subpheno_other}{dataframe for other subpheno}

\item{rho}{(optional) overlap between studies}
}
\value{
returns Combinedpheno dataframe that can be plugged into shaPRS_blend or shaPRS_composite
}
\description{
Convenience function to produce the combined phenotype estimate if we only have summary stats for phenoA and pheno B
}
