% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaPRS.R
\name{shaPRS}
\alias{shaPRS}
\title{Creates a new set of summary statistics}
\usage{
shaPRS(
  proximalLoc,
  adjunctLoc,
  outputLoc,
  rho = 0,
  discardAmbiguousSNPs = T,
  useProximalForMissing = T
)
}
\arguments{
\item{proximalLoc}{proximal LDPred formatted GWAS summary statistics table  that has header with the following columns: chr    pos    SNP    A1    A2    Freq1.Hapmap    b    se    p    N}

\item{adjunctLoc}{dataframe for adjunct dataset of the same signature}

\item{outputLoc}{the location of the output files}

\item{rho}{estimate of correlation between studies due to shared subjects. 0 for no overlap and 1 for complete overlap. default: 0. Obtain this from shaPRS_rho()}

\item{discardAmbiguousSNPs}{(optional) if ambiguous SNPs (G/C and A/T) should be discarded (default TRUE)}

\item{useProximalForMissing}{(optional) if SNPs missing from the adjunct data should be kept using the proximal data or not (default TRUE)}
}
\value{
returns object with two fields, (1) lFDRTable: a 3 column file with the following signature SNP lFDR Qval (2) hardThresholds list of SNPids that failed the heterogeneity test at each threshold
}
\description{
it performs both steps of the shaPRS method in a single call:
(1) Creates lFDR corrected Q-test statistics for each SNP (via shaPRS_adjust)
(2) produce summary statistics according to a continuous weighting scheme (via shaPRS_blend_overlap)
}
\examples{
proximalLoc <- system.file("extdata", "phenoA_sumstats", package = "shaPRS")
adjunctLoc <- system.file("extdata", "phenoB_sumstats", package = "shaPRS")
shaPRS(proximalLoc, adjunctLoc, tempfile())

}
